refactor
Refactors  SuntimesWarnings to no longer be an inner class (moved to .settings); adds accessors, adds listener.
